% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acf-package.R
\docType{package}
\name{Acf-package}
\alias{Acf}
\alias{Acf-package}
\title{Acf: Short and sweet function for plotting acf and pacf}
\description{
This is a wrapper for R functions acf and pacf which provides simultaneous plots for acf and pacf.
}
\keyword{internal}
