% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acf.R
\name{Acf}
\alias{Acf}
\alias{acf,Acf}
\title{Autocorrelation and partial autocorrelation functions for time series}
\usage{
Acf(x, lag.max = NULL, ...)
}
\arguments{
\item{x}{A univariate time series}

\item{lag.max}{(optional) Maximum lag at which to calculate the acf and pacf}

\item{...}{Further arguments to be passed to `acf` and `pacf`}
}
\description{
This function is a wrapper for R native functions `acf` and `pacf` which provides better plots.
}
\examples{
e <- rnorm(1000)
x <- filter(e, c(.5, .3), method="recursive")
Acf(x)
}
